refactor(pusher): honor backend ownership boundary
Move Pusher protocol and finalization helpers into utils and document the validated wire and durable ownership contracts. Verification: 146 focused tests passed; production Pyright passed; Black and diff checks passed.
