+++ backend/tests/unit/test_listen_finalization_cloud_tasks.py
@@ -19,8 +19,8 @@ from models.conversation_enums import ConversationStatus
 from routers.conversation_finalization import _parse_task_payload
 import routers.conversation_finalization as finalization_router
 import routers.pusher as pusher_router
-import routers.pusher_finalization as pusher_finalization
-import routers.pusher_protocol as pusher_protocol
+import utils.pusher_finalization as pusher_finalization
+import utils.pusher_protocol as pusher_protocol
 from utils.conversations import lifecycle as lifecycle_service
 from utils import app_integrations
 from utils import cloud_tasks
