+++ docs/doc/developer/backend/listen_pusher_pipeline.mdx
@@ -6,11 +6,25 @@ description: "Sequence diagrams for the /v4/listen WebSocket and Pusher processi
 
 # Listen + Pusher Pipeline — Sequence Diagrams
 
-> Last updated: 2026-07-15 (durable recording-session bindings, fenced late-content recovery, and ordered lifecycle envelopes)
+> Last updated: 2026-07-26 (bounded Pusher delivery, validated frames, and durable finalization ownership)
 >
 > These diagrams document the real behavior observed during E2E testing with
 > live services (backend, pusher, STT providers, embedding API). Update when the
 > pipeline changes.
+
+The backend-to-Pusher binary protocol accepts only opcodes `100` through `105`.
+Audio frames (`101`) require the four-byte opcode, an eight-byte timestamp, and
+PCM payload; JSON opcodes require an object payload, with transcript, finalization,
+and speaker-sample fields validated before they enter background queues. Malformed
+or unknown frames close the session with a protocol error instead of reaching
+application processing.
+
+Transcript, speaker-sample, audio-webhook, and private-cloud queues are bounded.
+Buffered audio also shares a per-session byte budget, so variable-size payloads
+cannot bypass item-count limits. A failed or cancelled backend-to-Pusher send
+retains the pending item for reconnect replay; buffers are cleared only after the
+send completes. Finalization opcode `104` carries the durable job id and dispatch
+generation, and Pusher claims that Firestore-owned job before processing.
 
 ## 1. Connection + Streaming + Transcription
